finish lab16

+++ src/main/java/com/application2/codeFellowship/Repository/PostRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post,Long> {
-    List<Post> findByUserId (Long id);
+    List<Post> findByUserId (Integer id);
 
 }
